Included stopOnError parameter

+++ vendors/migration.php
@@ -10,6 +10,11 @@ class Migration {
 	var $uses = array();
 
 	/**
+	 * Stop up/down on error
+	 */
+	var $stopOnError = true;
+
+	/**
 	 * DataSource link
 	 */
 	var $_db = null;
@@ -23,6 +28,11 @@ class Migration {
 	 * Fake CakeSchema
 	 */
 	var $__fakeSchema = null;
+
+	/**
+	 * Error
+	 */
+	var $__error = false;
 
 	/**
 	 * Constructor
@@ -70,6 +80,9 @@ class Migration {
      * Funçao de criaçao de tabela
      */
     function createTable($tableName, $columns, $indexes = array()) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$this->out('> ' . String::insert(__d('migrations', 'Creating table ":table"... ', true), array('table' => $tableName)), false);
 		$this->__fakeSchema->tables = array($tableName => $columns);
 		if (is_array($indexes) && !empty($indexes)) {
@@ -80,6 +93,7 @@ class Migration {
 			return true;
 		}
 		$this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -92,6 +106,9 @@ class Migration {
      * Funçao de excluir tabela
      */
     function dropTable($tableName) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$this->out('> ' . String::insert(__d('migrations', 'Dropping table ":table"... ', true), array('table' => $tableName)), false);
 		$this->__fakeSchema->tables = array($tableName => '');
 		if ($this->_db->execute($this->_db->dropSchema($this->__fakeSchema, $tableName))) {
@@ -99,6 +116,7 @@ class Migration {
 			return true;
 		}
 		$this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -106,6 +124,9 @@ class Migration {
      * Adicionar colunas
      */
     function addColumn($tableName, $columnName, $columnConfig = array()) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$columnConfig = array_merge(array('type' => 'integer'), $columnConfig);
 		$this->out('> ' . String::insert(__d('migrations', 'Creating column ":column"... ', true), array('column' => $columnName)), false);
 		if ($this->_db->execute($this->_db->alterSchema(array(
@@ -119,6 +140,7 @@ class Migration {
 			return true;
 		}
 		$this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -126,6 +148,9 @@ class Migration {
      * Remover colunas
      */
     function removeColumn($tableName, $columnName) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$this->out('> ' . String::insert(__d('migrations', 'Removing column ":column"... ', true), array('column' => $columnName)), false);
 		if ($this->_db->execute($this->_db->alterSchema(array(
 			$tableName => array(
@@ -138,6 +163,7 @@ class Migration {
 			return true;
 		}
 		$this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -145,11 +171,15 @@ class Migration {
      * Alterar colunas
      */
     function changeColumn($tableName, $columnName, $newColumnConfig = array(), $verbose = true) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$verbose && $this->out('> ' . String::insert(__d('migrations', 'Changing column ":column"... ', true), array('column' => $columnName)), false);
 		if ($this->_db->isInterfaceSupported('describe')) {
 			$describe = $this->_db->describe($tableName, true);
 			if (!isset($describe[$columnName])) {
 				$verbose &&  $this->out(__d('migrations', 'column not found.', true));
+				$this->__error = true;
 				return false;
 			}
 			$newColumnConfig = array_merge($describe[$columnName], $newColumnConfig);
@@ -165,6 +195,7 @@ class Migration {
 			return true;
 		}
 		$verbose && $this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -172,12 +203,16 @@ class Migration {
      * Renomear colunas
      */
     function renameColumn($tableName, $oldColumnName, $newColumnName) {
+		if ($this->stopOnError && $this->__error) {
+			return false;
+		}
 		$this->out('> ' . String::insert(__d('migrations', 'Renaming column ":old" to ":new"...', true), array('old' => $oldColumnName, 'new' => $newColumnName)), false);
 		if ($this->changeColumn($tableName, $oldColumnName, array('name' => $newColumnName), false)) {
 			$this->out('ok');
 			return true;
 		}
 		$this->out('nok');
+		$this->__error = true;
 		return false;
 	}
 
@@ -216,6 +251,7 @@ class Migration {
 	 * Execute Install and Uninstall methods
 	 */
 	function _exec($command, $callback) {
+		$this->__error = false;
 		if (!method_exists($this, $command)) {
 			$this->out(String::insert(__d('migrations', '> Method ":method" not implemented. Skipping...', true), array('method' => $command)));
 			return true;
@@ -226,13 +262,17 @@ class Migration {
 				return false;
 			}
 		}
-		$ok = true;
-		$this->_db->begin($this->__fakeSchema);
-		if (!$this->$command()) {
-			$this->_db->rollback($this->__fakeSchema);
-			$ok = false;
-		} else {
+		$ok = $this->_db->begin($this->__fakeSchema);
+		$this->$command();
+		if ($this->stopOnError) {
+			if ($this->__error) {
+				$ok = false;
+			}
+		}
+		if ($ok) {
 			$this->_db->commit($this->__fakeSchema);
+		} else {
+			$this->_db->rollback($this->__fakeSchema);
 		}
 		$method = 'after' . $callback;
 		if (method_exists($this, $method)) {
